+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.1.0
 
+- Added the team namespace setting
 - Fixed a bug in determining the repository owner
 
 ### 1.0.0